+++ benchmarks/decode_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"testing"
 
+	"github.com/bytedance/sonic"
 	gojay "github.com/francoispqt/gojay"
 	gojson "github.com/goccy/go-json"
 	jsoniter "github.com/json-iterator/go"
@@ -28,6 +29,26 @@ func Benchmark_Decode_SmallStruct_Unmarshal_FastJson(b *testing.B) {
 	for n := 0; n < b.N; n++ {
 		var p fastjson.Parser
 		if _, err := p.Parse(smallFixture); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_SmallStruct_Unmarshal_SegmentioJson(b *testing.B) {
+	b.ReportAllocs()
+	for n := 0; n < b.N; n++ {
+		result := SmallPayload{}
+		if err := segmentiojson.Unmarshal(SmallFixture, &result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_SmallStruct_Unmarshal_Sonic(b *testing.B) {
+	b.ReportAllocs()
+	for n := 0; n < b.N; n++ {
+		result := SmallPayload{}
+		if err := sonic.Unmarshal(SmallFixture, &result); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -58,16 +79,6 @@ func Benchmark_Decode_SmallStruct_Unmarshal_GoJayUnsafe(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		result := SmallPayload{}
 		if err := gojay.Unsafe.UnmarshalJSONObject(SmallFixture, &result); err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func Benchmark_Decode_SmallStruct_Unmarshal_SegmentioJson(b *testing.B) {
-	b.ReportAllocs()
-	for n := 0; n < b.N; n++ {
-		result := SmallPayload{}
-		if err := segmentiojson.Unmarshal(SmallFixture, &result); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -105,6 +116,18 @@ func Benchmark_Decode_SmallStruct_Stream_EncodingJson(b *testing.B) {
 	}
 }
 
+func Benchmark_Decode_SmallStruct_Stream_SegmentioJson(b *testing.B) {
+	b.ReportAllocs()
+	reader := bytes.NewReader(SmallFixture)
+	for i := 0; i < b.N; i++ {
+		result := SmallPayload{}
+		reader.Reset(SmallFixture)
+		if err := segmentiojson.NewDecoder(reader).Decode(&result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Decode_SmallStruct_Stream_JsonIter(b *testing.B) {
 	b.ReportAllocs()
 	reader := bytes.NewReader(SmallFixture)
@@ -124,18 +147,6 @@ func Benchmark_Decode_SmallStruct_Stream_GoJay(b *testing.B) {
 		reader.Reset(SmallFixture)
 		result := SmallPayload{}
 		if err := gojay.NewDecoder(reader).DecodeObject(&result); err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func Benchmark_Decode_SmallStruct_Stream_SegmentioJson(b *testing.B) {
-	b.ReportAllocs()
-	reader := bytes.NewReader(SmallFixture)
-	for i := 0; i < b.N; i++ {
-		result := SmallPayload{}
-		reader.Reset(SmallFixture)
-		if err := segmentiojson.NewDecoder(reader).Decode(&result); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -174,6 +185,26 @@ func Benchmark_Decode_MediumStruct_Unmarshal_FastJson(b *testing.B) {
 	}
 }
 
+func Benchmark_Decode_MediumStruct_Unmarshal_Sonic(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		result := MediumPayload{}
+		if err := sonic.Unmarshal(MediumFixture, &result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_MediumStruct_Unmarshal_SegmentioJson(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		result := MediumPayload{}
+		if err := segmentiojson.Unmarshal(MediumFixture, &result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Decode_MediumStruct_Unmarshal_JsonIter(b *testing.B) {
 	b.ReportAllocs()
 	for n := 0; n < b.N; n++ {
@@ -199,16 +230,6 @@ func Benchmark_Decode_MediumStruct_Unmarshal_GoJayUnsafe(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		result := MediumPayload{}
 		if err := gojay.Unsafe.UnmarshalJSONObject(MediumFixture, &result); err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func Benchmark_Decode_MediumStruct_Unmarshal_SegmentioJson(b *testing.B) {
-	b.ReportAllocs()
-	for i := 0; i < b.N; i++ {
-		result := MediumPayload{}
-		if err := segmentiojson.Unmarshal(MediumFixture, &result); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -246,6 +267,18 @@ func Benchmark_Decode_MediumStruct_Stream_EncodingJson(b *testing.B) {
 	}
 }
 
+func Benchmark_Decode_MediumStruct_Stream_SegmentioJson(b *testing.B) {
+	b.ReportAllocs()
+	reader := bytes.NewReader(MediumFixture)
+	for n := 0; n < b.N; n++ {
+		reader.Reset(MediumFixture)
+		result := MediumPayload{}
+		if err := segmentiojson.NewDecoder(reader).Decode(&result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Decode_MediumStruct_Stream_JsonIter(b *testing.B) {
 	b.ReportAllocs()
 	reader := bytes.NewReader(MediumFixture)
@@ -265,18 +298,6 @@ func Benchmark_Decode_MediumStruct_Stream_GoJay(b *testing.B) {
 		reader.Reset(MediumFixture)
 		result := MediumPayload{}
 		if err := gojay.NewDecoder(reader).DecodeObject(&result); err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func Benchmark_Decode_MediumStruct_Stream_SegmentioJson(b *testing.B) {
-	b.ReportAllocs()
-	reader := bytes.NewReader(MediumFixture)
-	for n := 0; n < b.N; n++ {
-		reader.Reset(MediumFixture)
-		result := MediumPayload{}
-		if err := segmentiojson.NewDecoder(reader).Decode(&result); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -315,6 +336,26 @@ func Benchmark_Decode_LargeStruct_Unmarshal_FastJson(b *testing.B) {
 	}
 }
 
+func Benchmark_Decode_LargeStruct_Unmarshal_Sonic(b *testing.B) {
+	b.ReportAllocs()
+	for n := 0; n < b.N; n++ {
+		result := LargePayload{}
+		if err := sonic.Unmarshal(LargeFixture, &result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_LargeStruct_Unmarshal_SegmentioJson(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		result := LargePayload{}
+		if err := segmentiojson.Unmarshal(LargeFixture, &result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Decode_LargeStruct_Unmarshal_JsonIter(b *testing.B) {
 	b.ReportAllocs()
 	for n := 0; n < b.N; n++ {
@@ -345,16 +386,6 @@ func Benchmark_Decode_LargeStruct_Unmarshal_GoJayUnsafe(b *testing.B) {
 	}
 }
 
-func Benchmark_Decode_LargeStruct_Unmarshal_SegmentioJson(b *testing.B) {
-	b.ReportAllocs()
-	for i := 0; i < b.N; i++ {
-		result := LargePayload{}
-		if err := segmentiojson.Unmarshal(LargeFixture, &result); err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
 func Benchmark_Decode_LargeStruct_Unmarshal_GoJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
@@ -375,6 +406,34 @@ func Benchmark_Decode_LargeStruct_Unmarshal_GoJsonNoEscape(b *testing.B) {
 	}
 }
 
+func Benchmark_Decode_LargeStruct_Unmarshal_GoJsonFirstWinMode(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		result := LargePayload{}
+		if err := gojson.UnmarshalWithOption(
+			LargeFixture,
+			&result,
+			gojson.DecodeFieldPriorityFirstWin(),
+		); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_LargeStruct_Unmarshal_GoJsonNoEscapeFirstWinMode(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		result := LargePayload{}
+		if err := gojson.UnmarshalNoEscape(
+			LargeFixture,
+			&result,
+			gojson.DecodeFieldPriorityFirstWin(),
+		); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Decode_LargeStruct_Stream_EncodingJson(b *testing.B) {
 	b.ReportAllocs()
 	reader := bytes.NewReader(LargeFixture)
@@ -382,6 +441,18 @@ func Benchmark_Decode_LargeStruct_Stream_EncodingJson(b *testing.B) {
 		result := LargePayload{}
 		reader.Reset(LargeFixture)
 		if err := json.NewDecoder(reader).Decode(&result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_LargeStruct_Stream_SegmentioJson(b *testing.B) {
+	b.ReportAllocs()
+	reader := bytes.NewReader(LargeFixture)
+	for i := 0; i < b.N; i++ {
+		result := LargePayload{}
+		reader.Reset(LargeFixture)
+		if err := segmentiojson.NewDecoder(reader).Decode(&result); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -411,18 +482,6 @@ func Benchmark_Decode_LargeStruct_Stream_GoJay(b *testing.B) {
 	}
 }
 
-func Benchmark_Decode_LargeStruct_Stream_SegmentioJson(b *testing.B) {
-	b.ReportAllocs()
-	reader := bytes.NewReader(LargeFixture)
-	for i := 0; i < b.N; i++ {
-		result := LargePayload{}
-		reader.Reset(LargeFixture)
-		if err := segmentiojson.NewDecoder(reader).Decode(&result); err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
 func Benchmark_Decode_LargeStruct_Stream_GoJson(b *testing.B) {
 	b.ReportAllocs()
 	reader := bytes.NewReader(LargeFixture)
@@ -430,6 +489,21 @@ func Benchmark_Decode_LargeStruct_Stream_GoJson(b *testing.B) {
 		result := LargePayload{}
 		reader.Reset(LargeFixture)
 		if err := gojson.NewDecoder(reader).Decode(&result); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Decode_LargeStruct_Stream_GoJsonFirstWinMode(b *testing.B) {
+	b.ReportAllocs()
+	reader := bytes.NewReader(LargeFixture)
+	for i := 0; i < b.N; i++ {
+		result := LargePayload{}
+		reader.Reset(LargeFixture)
+		if err := gojson.NewDecoder(reader).DecodeWithOption(
+			&result,
+			gojson.DecodeFieldPriorityFirstWin(),
+		); err != nil {
 			b.Fatal(err)
 		}
 	}