+++ benchmarks/encode_test.go
@@ -5,9 +5,11 @@ import (
 	"encoding/json"
 	"testing"
 
+	"github.com/bytedance/sonic"
 	gojay "github.com/francoispqt/gojay"
 	gojson "github.com/goccy/go-json"
 	jsoniter "github.com/json-iterator/go"
+	"github.com/pquerna/ffjson/ffjson"
 	segmentiojson "github.com/segmentio/encoding/json"
 	"github.com/wI2L/jettison"
 )
@@ -16,6 +18,15 @@ func Benchmark_Encode_SmallStruct_EncodingJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		if _, err := json.Marshal(NewSmallPayload()); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_SmallStruct_FFJson(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := ffjson.Marshal(NewSmallPayloadFFJson()); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -67,6 +78,25 @@ func Benchmark_Encode_SmallStruct_SegmentioJson(b *testing.B) {
 	}
 }
 
+func Benchmark_Encode_SmallStruct_Sonic(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := sonic.Marshal(NewSmallPayload()); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_SmallStruct_GoJsonColored(b *testing.B) {
+	colorOpt := gojson.Colorize(gojson.DefaultColorScheme)
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := gojson.MarshalWithOption(NewSmallPayload(), colorOpt); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Encode_SmallStruct_GoJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
@@ -90,6 +120,16 @@ func Benchmark_Encode_SmallStructCached_EncodingJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		if _, err := json.Marshal(cached); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_SmallStructCached_FFJson(b *testing.B) {
+	cached := NewSmallPayloadFFJson()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := ffjson.Marshal(cached); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -141,6 +181,27 @@ func Benchmark_Encode_SmallStructCached_SegmentioJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		if _, err := segmentiojson.Marshal(cached); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_SmallStructCached_Sonic(b *testing.B) {
+	cached := NewSmallPayload()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := sonic.Marshal(cached); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_SmallStructCached_GoJsonColored(b *testing.B) {
+	cached := NewSmallPayload()
+	colorOpt := gojson.Colorize(gojson.DefaultColorScheme)
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := gojson.MarshalWithOption(cached, colorOpt); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -216,6 +277,25 @@ func Benchmark_Encode_MediumStruct_SegmentioJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		if _, err := segmentiojson.Marshal(NewMediumPayload()); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_MediumStruct_Sonic(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := sonic.Marshal(NewMediumPayload()); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_MediumStruct_GoJsonColored(b *testing.B) {
+	colorOpt := gojson.Colorize(gojson.DefaultColorScheme)
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := gojson.MarshalWithOption(NewMediumPayload(), colorOpt); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -300,6 +380,27 @@ func Benchmark_Encode_MediumStructCached_SegmentioJson(b *testing.B) {
 	}
 }
 
+func Benchmark_Encode_MediumStructCached_Sonic(b *testing.B) {
+	cached := NewMediumPayload()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := sonic.Marshal(cached); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_MediumStructCached_GoJsonColored(b *testing.B) {
+	cached := NewMediumPayload()
+	colorOpt := gojson.Colorize(gojson.DefaultColorScheme)
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := gojson.MarshalWithOption(cached, colorOpt); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func Benchmark_Encode_MediumStructCached_GoJson(b *testing.B) {
 	cached := NewMediumPayload()
 	b.ReportAllocs()
@@ -370,6 +471,25 @@ func Benchmark_Encode_LargeStruct_SegmentioJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		if _, err := segmentiojson.Marshal(NewLargePayload()); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_LargeStruct_Sonic(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := sonic.Marshal(NewLargePayload()); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_LargeStruct_GoJsonColored(b *testing.B) {
+	colorOpt := gojson.Colorize(gojson.DefaultColorScheme)
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := gojson.MarshalWithOption(NewLargePayload(), colorOpt); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -449,6 +569,27 @@ func Benchmark_Encode_LargeStructCached_SegmentioJson(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		if _, err := segmentiojson.Marshal(cached); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_LargeStructCached_Sonic(b *testing.B) {
+	cached := NewLargePayload()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := sonic.Marshal(cached); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func Benchmark_Encode_LargeStructCached_GoJsonColored(b *testing.B) {
+	cached := NewLargePayload()
+	colorOpt := gojson.Colorize(gojson.DefaultColorScheme)
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if _, err := gojson.MarshalWithOption(cached, colorOpt); err != nil {
 			b.Fatal(err)
 		}
 	}