+++ internal/cmd/generator/main.go
@@ -4,6 +4,9 @@ import (
 	"bytes"
 	"fmt"
 	"go/format"
+	"go/parser"
+	"go/printer"
+	"go/token"
 	"io/ioutil"
 	"path/filepath"
 	"runtime"
@@ -45,7 +48,7 @@ var opTypeStrings = [{{ .OpLen }}]string{
 {{- end }}
 }
 
-type OpType int
+type OpType uint16
 
 const (
 {{- range $index, $type := .OpTypes }}
@@ -100,7 +103,7 @@ func (t OpType) HeadToPtrHead() OpType {
   }
   suffix := "PtrHead"+t.String()[idx+len("Head"):]
 
-  const toPtrOffset = 3
+  const toPtrOffset = 2
   if strings.Contains(OpType(int(t) + toPtrOffset).String(), suffix) {
     return OpType(int(t) + toPtrOffset)
   }
@@ -116,14 +119,6 @@ func (t OpType) HeadToOmitEmptyHead() OpType {
   return t
 }
 
-func (t OpType) HeadToStringTagHead() OpType {
-  const toStringTagOffset = 2
-  if strings.Contains(OpType(int(t) + toStringTagOffset).String(), "StringTag") {
-    return OpType(int(t) + toStringTagOffset)
-  }
-  return t
-}
-
 func (t OpType) PtrHeadToHead() OpType {
   idx := strings.Index(t.String(), "Ptr")
   if idx == -1 {
@@ -131,7 +126,7 @@ func (t OpType) PtrHeadToHead() OpType {
   }
   suffix := t.String()[idx+len("Ptr"):]
 
-  const toPtrOffset = 3
+  const toPtrOffset = 2
   if strings.Contains(OpType(int(t) - toPtrOffset).String(), suffix) {
     return OpType(int(t) - toPtrOffset)
   }
@@ -144,10 +139,10 @@ func (t OpType) FieldToEnd() OpType {
     return t
   }
   suffix := t.String()[idx+len("Field"):]
-  if suffix == "" || suffix == "OmitEmpty" || suffix == "StringTag" {
+  if suffix == "" || suffix == "OmitEmpty" {
     return t
   }
-  const toEndOffset = 3
+  const toEndOffset = 2
   if strings.Contains(OpType(int(t) + toEndOffset).String(), "End"+suffix) {
     return OpType(int(t) + toEndOffset)
   }
@@ -161,15 +156,6 @@ func (t OpType) FieldToOmitEmptyField() OpType {
   }
   return t
 }
-
-func (t OpType) FieldToStringTagField() OpType {
-  const toStringTagOffset = 2
-  if strings.Contains(OpType(int(t) + toStringTagOffset).String(), "StringTag") {
-    return OpType(int(t) + toStringTagOffset)
-  }
-  return t
-}
-
 `)
 	if err != nil {
 		return err
@@ -191,9 +177,10 @@ func (t OpType) FieldToStringTagField() OpType {
 	primitiveTypes := []string{
 		"int", "uint", "float32", "float64", "bool", "string", "bytes", "number",
 		"array", "map", "slice", "struct", "MarshalJSON", "MarshalText",
-		"intString", "uintString",
+		"intString", "uintString", "float32String", "float64String", "boolString", "stringString", "numberString",
 		"intPtr", "uintPtr", "float32Ptr", "float64Ptr", "boolPtr", "stringPtr", "bytesPtr", "numberPtr",
 		"arrayPtr", "mapPtr", "slicePtr", "marshalJSONPtr", "marshalTextPtr", "interfacePtr",
+		"intPtrString", "uintPtrString", "float32PtrString", "float64PtrString", "boolPtrString", "stringPtrString", "numberPtrString",
 	}
 	primitiveTypesUpper := []string{}
 	for _, typ := range primitiveTypes {
@@ -213,6 +200,7 @@ func (t OpType) FieldToStringTagField() OpType {
 		createOpType("Recursive", "Op"),
 		createOpType("RecursivePtr", "Op"),
 		createOpType("RecursiveEnd", "Op"),
+		createOpType("InterfaceEnd", "Op"),
 		createOpType("StructAnonymousEnd", "StructEnd"),
 	}
 	for _, typ := range primitiveTypesUpper {
@@ -221,7 +209,7 @@ func (t OpType) FieldToStringTagField() OpType {
 	}
 	for _, typ := range append(primitiveTypesUpper, "") {
 		for _, ptrOrNot := range []string{"", "Ptr"} {
-			for _, opt := range []string{"", "OmitEmpty", "StringTag"} {
+			for _, opt := range []string{"", "OmitEmpty"} {
 				ptrOrNot := ptrOrNot
 				opt := opt
 				typ := typ
@@ -240,7 +228,7 @@ func (t OpType) FieldToStringTagField() OpType {
 		}
 	}
 	for _, typ := range append(primitiveTypesUpper, "") {
-		for _, opt := range []string{"", "OmitEmpty", "StringTag"} {
+		for _, opt := range []string{"", "OmitEmpty"} {
 			opt := opt
 			typ := typ
 
@@ -254,7 +242,7 @@ func (t OpType) FieldToStringTagField() OpType {
 				Code: "StructField",
 			})
 		}
-		for _, opt := range []string{"", "OmitEmpty", "StringTag"} {
+		for _, opt := range []string{"", "OmitEmpty"} {
 			opt := opt
 			typ := typ
 
@@ -289,6 +277,32 @@ func (t OpType) FieldToStringTagField() OpType {
 	return ioutil.WriteFile(path, buf, 0644)
 }
 
+func generateVM() error {
+	file, err := ioutil.ReadFile("vm.go.tmpl")
+	if err != nil {
+		return err
+	}
+	fset := token.NewFileSet()
+	f, err := parser.ParseFile(fset, "", string(file), parser.ParseComments)
+	if err != nil {
+		return err
+	}
+	for _, pkg := range []string{"vm", "vm_indent", "vm_color", "vm_color_indent"} {
+		f.Name.Name = pkg
+		var buf bytes.Buffer
+		printer.Fprint(&buf, fset, f)
+		path := filepath.Join(repoRoot(), "internal", "encoder", pkg, "vm.go")
+		source, err := format.Source(buf.Bytes())
+		if err != nil {
+			return err
+		}
+		if err := ioutil.WriteFile(path, source, 0644); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func repoRoot() string {
 	_, file, _, _ := runtime.Caller(0)
 	relativePathFromRepoRoot := filepath.Join("internal", "cmd", "generator")
@@ -297,6 +311,9 @@ func repoRoot() string {
 
 //go:generate go run main.go
 func main() {
+	if err := generateVM(); err != nil {
+		panic(err)
+	}
 	if err := _main(); err != nil {
 		panic(err)
 	}